[General] Clean Match entity [Profile] UI fixes [General] Get Shots from server on return from background

+++ Goles/Utils/Utils.m
@@ -94,7 +94,7 @@
 
 + (UIView *)setNavigationBarTitle:(NSString *)title andSubtitle:(NSString *)subtitle forLoadingMode:(BOOL)loadingMode forMaximumLenght:(NSNumber *)lenghtView{
     
-    int lengthOfString = [title length] ;
+    NSInteger lengthOfString = [title length] ;
     
     if (lenghtView > 0)
         lengthOfString = [lenghtView integerValue];
